formatting & bugfix

+++ 7.Mitm/hyperfox/hyperfox-update.sh
@@ -3,6 +3,28 @@
 bold=$(tput bold)
 normal=$(tput sgr0)
 
+GITREPO=https://github.com/malfunkt/hyperfox.git
+BRANCH=master
+GITREPOROOT=/opt/ITSEC/7.Mitm/hyperfox/malfunkt/hyperfox
+GITCONFDIR=/opt/ITSEC/7.Mitm/hyperfox/malfunkt/hyperfox/.git
+GITCLONEDIR=/opt/ITSEC/7.Mitm/hyperfox/malfunkt
+BINDIR=/usr/local/bin
+DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/7.Mitm
+DSKTPFLSDEST=/home/$USER/.local/share/applications/7.Mitm
+DSKTPFL1=hyperfox.desktop
+DSKTPFL1=arpfox.desktop
+GITRESET () {
+	git clean -f
+	git fetch origin
+	git reset --hard origin/$BRANCH
+	git pull
+}
+GITSBMDLINIT () {
+	git submodule init
+	git submodule update --recursive
+	sudo updatedb && sudo ldconfig
+}
+
 echo "${bold}
  _   ___   ______  _____ ____  _____ _____  __
 | | | \ \ / /  _ \| ____|  _ \|  ___/ _ \ \/ /
@@ -17,55 +39,59 @@ echo "${bold}
  / ___ \|  _ <|  __/|  _|| |_| /  \ 
 /_/   \_\_| \_\_|   |_|   \___/_/\_\
            
+UPDATE
 ${normal}"
 
-GITREPOROOT=/opt/ITSEC/7.Mitm/hyperfox/malfunkt/hyperfox
-GITREPOGITFILE=$GITREPOROOT/.git
-DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/7.Mitm
-DSKTPFLSDEST=/home/$USER/.local/share/applications/7.Mitm
-DSKTPFL1=hyperfox.desktop
-DSKTPFL1=arpfox.desktop
-
-
-if [ ! -d $GITREPOGITFILE ]
+if [ ! -d $GITCONFDIR ]
 
 then
 
-mkdir -p /opt/ITSEC/7.Mitm/hyperfox/malfunkt
-cd /opt/ITSEC/7.Mitm/hyperfox/malfunkt
-git clone https://github.com/malfunkt/hyperfox.git
+mkdir -p $GITCLONEDIR
+cd $GITCLONEDIR
+git clone -b $BRANCH $GITREPO
 
 else
 
-echo "repo exists"
+echo "${bold}REPO EXISTS, skip clone...${normal}"
 
 fi
 
 cd $GITREPOROOT
 
-if git diff-index --quiet HEAD --; then
-    echo "UP TO DATE"
+if git checkout $BRANCH &&
+    git fetch origin $BRANCH &&
+    [ `git rev-list HEAD...origin/$BRANCH --count` != 0 ] &&
+    git merge origin/$BRANCH
+then
+    
+cd $GITREPOROOT
 
-else
-
-sudo rm -f $DSKTPFLSDEST/$DSKTPFL1
-sudo rm -f $DSKTPFLSDEST/$DSKTPFL2
-sudo rm -f /usr/local/bin/hyperfox
-sudo rm -f /usr/local/bin/arpfox
-
+GITRESET
+GITSBMDLINIT
 #cd /opt/ITSEC/7.Mitm/hyperfox/malfunkt/hyperfox
-curl -sL 'https://raw.githubusercontent.com/malfunkt/hyperfox/master/install.sh' | sh
-curl -sL 'https://raw.githubusercontent.com/malfunkt/arpfox/master/install.sh' | sh
+curl -sL 'https://raw.githubusercontent.com/malfunkt/hyperfox/$BRANCH/install.sh' | sh
+curl -sL 'https://raw.githubusercontent.com/malfunkt/arpfox/$BRANCH/install.sh' | sh
 
 sudo updatedb 
 sudo ldconfig
 
 #sudo ln -s /opt/ITSEC/7.Mitm/hyperfox/malfunkt/hyperfox/hyperfox.sh /usr/local/bin/hyperfox
-rm -f $DSKTPFLSDEST/$DSKTPFL1
-rm -f $DSKTPFLSDEST/$DSKTPFL2
+sudo rm -f /usr/local/bin/hyperfox
+sudo rm -f /usr/local/bin/arpfox
 mkdir -p $DSKTPFLSDEST
+sudo rm -f $DSKTPFLSDEST/$DSKTPFL1
+sudo rm -f $DSKTPFLSDEST/$DSKTPFL2
 cp $DSKTPFLS/$DSKTPFL1 $DSKTPFLSDEST/$DSKTPFL1
 cp $DSKTPFLS/$DSKTPFL2 $DSKTPFLSDEST/$DSKTPFL2
 
+echo "${bold}
+UPDATED
+${normal}"
 
+else
+
+echo "${bold}
+UP TO DATE
+${normal}"
+	
 fi
