+++ 7.Mitm/sslsplit/sslsplit-update.sh
@@ -3,61 +3,82 @@
 bold=$(tput bold)
 normal=$(tput sgr0)
 
+GITREPO=https://github.com/droe/sslsplit.git
+BRANCH=master
+GITREPOROOT=/opt/ITSEC/7.Mitm/sslsplit/droe/sslsplit
+GITCONFDIR=/opt/ITSEC/7.Mitm/sslsplit/droe/sslsplit/.git
+GITCLONEDIR=/opt/ITSEC/7.Mitm/sslsplit/droe
+DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/7.Mitm
+DSKTPFLSDEST=/home/$USER/.local/share/applications/7.Mitm
+DSKTPFL=sslsplit.desktop
+
+GITRESET () {
+	git clean -f
+	git fetch origin
+	git reset --hard origin/$BRANCH
+	git pull
+}
+GITSBMDLINIT () {
+	git submodule init
+	git submodule update --recursive
+	sudo updatedb && sudo ldconfig
+}
+
 echo "${bold}
  ____ ____  _     ____  ____  _     ___ _____ 
 / ___/ ___|| |   / ___||  _ \| |   |_ _|_   _|
 \___ \___ \| |   \___ \| |_) | |    | |  | |  
  ___) |__) | |___ ___) |  __/| |___ | |  | |  
 |____/____/|_____|____/|_|   |_____|___| |_|  
-             
+            
+UPDATE 
 ${normal}"
 
-
-GITREPOROOT=/opt/ITSEC/7.Mitm/sslsplit/droe/sslsplit
-GITREPOGITFILE=$GITREPOROOT/.git
-DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/7.Mitm
-DSKTPFLSDEST=/home/$USER/.local/share/applications/7.Mitm
-DSKTPFL=sslsplit.desktop
-
-if [ ! -d $GITREPOGITFILE ]
+if [ ! -d $GITCONFDIR ]
 
 then
 
-mkdir -p /opt/ITSEC/7.Mitm/sslsplit/droe
-cd /opt/ITSEC/7.Mitm/sslsplit/droe
-git clone https://github.com/droe/sslsplit.git
+mkdir -p $GITCLONEDIR
+cd $GITCLONEDIR
+git clone -b $BRANCH $GITREPO
 
 else
 
-echo "repo exists"
+echo "${bold}REPO EXISTS, skip clone...${normal}"
 
 fi
 
 cd $GITREPOROOT
 
-if git diff-index --quiet HEAD --; then
-    echo "UP TO DATE"
-
-else
-
-
-
+if git checkout $BRANCH &&
+    git fetch origin $BRANCH &&
+    [ `git rev-list HEAD...origin/$BRANCH --count` != 0 ] &&
+    git merge origin/$BRANCH
+then
+    
 cd $GITREPOROOT
 sudo make uninstall
 sudo updatedb
 sudo ldconfig
 make clean
-git clean -f 
-git fetch origin
-git reset --hard origin/master
-git pull
-git submodule init
-git submodule update --recursive
+GITRESET
+GITSBMDLINIT
 make -j 4
 sudo make install
 rm -f $DSKTPFLSDEST/$DSKTPFL
 mkdir -p $DSKTPFLSDEST
 cp $DSKTPFLS/$DSKTPFL $DSKTPFLSDEST/$DSKTPFL
 
+echo "${bold}
+UPDATED
+${normal}"
+
+else
+
+echo "${bold}
+UP TO DATE
+${normal}"
+	
 fi
+
