+++ 8.Tunnel/dns2tcp/dns2tcp-update.sh
@@ -3,62 +3,82 @@
 bold=$(tput bold)
 normal=$(tput sgr0)
 
+GITREPO=https://github.com/risent/dns2tcp.git
+BRANCH=master
+GITREPOROOT=/opt/ITSEC/8.Tunnel/dns2tcp/risent/dns2tcp
+GITCONFDIR=/opt/ITSEC/8.Tunnel/dns2tcp/risent/dns2tcp/.git
+GITCLONEDIR=/opt/ITSEC/8.Tunnel/dns2tcp/risent
+EXECUTEABLE1=dns2tcp
+EXECUTEABLE2=dns2tcp
+BINDIR=/usr/local/bin
+DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/8.Tunnel
+DSKTPFLSDEST=/home/$USER/.local/share/applications/8.Tunnel
+DSKTPFL=dns2tcp.desktop
+GITRESET () {
+	git clean -f
+	git fetch origin
+	git reset --hard origin/$BRANCH
+	git pull
+}
+GITSBMDLINIT () {
+	git submodule init
+	git submodule update --recursive
+	sudo updatedb && sudo ldconfig
+}
+
 echo "${bold}
  ____  _   _ ____ ____ _____ ____ ____  
 |  _ \| \ | / ___|___ \_   _/ ___|  _ \ 
 | | | |  \| \___ \ __) || || |   | |_) |
 | |_| | |\  |___) / __/ | || |___|  __/ 
 |____/|_| \_|____/_____||_| \____|_|    
-                                             
+              
+UPDATE                               
 ${normal}"
 
-GITREPOROOT=/opt/ITSEC/8.Tunnel/dns2tcp/risent/dns2tcp
-GITREPOGITFILE=$GITREPOROOT/.git
-
-DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/8.Tunnel
-DSKTPFLSDEST=/home/$USER/.local/share/applications/8.Tunnel
-DSKTPFL=dns2tcp.desktop
-
-if [ ! -d $GITREPOGITFILE ]
+if [ ! -d $GITCONFDIR ]
 
 then
 
-
-mkdir -p /opt/ITSEC/8.Tunnel/dns2tcp/risent
-cd /opt/ITSEC/8.Tunnel/dns2tcp/risent
-git clone https://github.com/risent/dns2tcp.git
+mkdir -p $GITCLONEDIR
+cd $GITCLONEDIR
+git clone -b $BRANCH $GITREPO
 
 else
 
-echo "repo exists"
+echo "${bold}REPO EXISTS, skip clone...${normal}"
 
 fi
 
 cd $GITREPOROOT
 
-if git diff-index --quiet HEAD --; then
-    echo "UP TO DATE"
-
-else
-
-
-
+if git checkout $BRANCH &&
+    git fetch origin $BRANCH &&
+    [ `git rev-list HEAD...origin/$BRANCH --count` != 0 ] &&
+    git merge origin/$BRANCH
+then
+    
 cd $GITREPOROOT
-git clean -f
-git fetch origin
-git reset --hard origin/master
-git pull
-git submodule init
-git submodule update --recursive
-#
+GITRESET
+GITSBMDLINIT
 go build
 
-chmod +x dns2tcp
-sudo rm -f /usr/local/bin/dns2tcp
-sudo ln -s $GITREPOROOT/dns2tcp /usr/local/bin/dns2tcp
+chmod +x $EXECUTEABLE1
+sudo rm -f $BINDIR/$EXECUTEABLE2
+sudo ln -s $GITREPOROOT/$EXECUTEABLE1 $BINDIR/$EXECUTEABLE2
 
 rm -f $DSKTPFLSDEST/$DSKTPFL
 mkdir -p $DSKTPFLSDEST
 cp $DSKTPFLS/$DSKTPFL $DSKTPFLSDEST/$DSKTPFL
 
+echo "${bold}
+UPDATED
+${normal}"
+
+else
+
+echo "${bold}
+UP TO DATE
+${normal}"
+	
 fi
