+++ 8.Tunnel/iodine/iodine-update.sh
@@ -3,60 +3,78 @@
 bold=$(tput bold)
 normal=$(tput sgr0)
 
+GITREPO=https://github.com/yarrick/iodine.git
+BRANCH=master
+GITREPOROOT=/opt/ITSEC/8.Tunnel/iodine/yarrick/iodine
+GITCONFDIR=/opt/ITSEC/8.Tunnel/iodine/yarrick/iodine/.git
+GITCLONEDIR=/opt/ITSEC/8.Tunnel/iodine/yarrick/iodine
+DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/8.Tunnel
+DSKTPFLSDEST=/home/$USER/.local/share/applications/8.Tunnel
+DSKTPFL=iodine.desktop
+GITRESET () {
+	git clean -f
+	git fetch origin
+	git reset --hard origin/$BRANCH
+	git pull
+}
+GITSBMDLINIT () {
+	git submodule init
+	git submodule update --recursive
+	sudo updatedb && sudo ldconfig
+}
+
 echo "${bold}
  ___ ___  ____ ___ _   _ _____ 
 |_ _/ _ \|  _ \_ _| \ | | ____|
  | | | | | | | | ||  \| |  _|  
  | | |_| | |_| | || |\  | |___ 
 |___\___/|____/___|_| \_|_____|
-               
+            
+UPDATE   
 ${normal}"
 
-GITREPOROOT=/opt/ITSEC/8.Tunnel/iodine/yarrick/iodine
-GITREPOGITFILE=$GITREPOROOT/.git
-
-DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/8.Tunnel
-DSKTPFLSDEST=/home/$USER/.local/share/applications/8.Tunnel
-DSKTPFL=iodine.desktop
-
-if [ ! -d $GITREPOGITFILE ]
+if [ ! -d $GITCONFDIR ]
 
 then
 
-
-mkdir -p /opt/ITSEC/8.Tunnel/iodine/yarrick
-cd /opt/ITSEC/8.Tunnel/iodine/yarrick
-git clone https://github.com/yarrick/iodine.git
+mkdir -p $GITCLONEDIR
+cd $GITCLONEDIR
+git clone -b $BRANCH $GITREPO
 
 else
 
-echo "repo exists"
+echo "${bold}REPO EXISTS, skip clone...${normal}"
 
 fi
 
 cd $GITREPOROOT
 
-if git diff-index --quiet HEAD --; then
-    echo "UP TO DATE"
-
-else
-
-
+if git checkout $BRANCH &&
+    git fetch origin $BRANCH &&
+    [ `git rev-list HEAD...origin/$BRANCH --count` != 0 ] &&
+    git merge origin/$BRANCH
+then
+    
 cd $GITREPOROOT
 sudo make uninstall
 make clean
-git clean -f
-git fetch origin
-git reset --hard origin/master
-git pull
-git submodule init 
-git submodule update --recursive
-#
+GITRESET
+GITSBMDLINIT
 make -j 4
 sudo make install
 rm -f $DSKTPFLSDEST/$DSKTPFL
 mkdir -p $DSKTPFLSDEST
 cp $DSKTPFLS/$DSKTPFL $DSKTPFLSDEST/$DSKTPFL
 
+echo "${bold}
+UPDATED
+${normal}"
+
+else
+
+echo "${bold}
+UP TO DATE
+${normal}"
+	
 fi
