+++ 8.Tunnel/sslh/sslh-update.sh
@@ -3,6 +3,26 @@
 bold=$(tput bold)
 normal=$(tput sgr0)
 
+GITREPO=https://github.com/yrutschle/sslh.git
+BRANCH=master
+GITREPOROOT=/opt/ITSEC/8.Tunnel/sslh/yrutschle/sslh
+GITCONFDIR=/opt/ITSEC/8.Tunnel/sslh/yrutschle/sslh/.git
+GITCLONEDIR=/opt/ITSEC/8.Tunnel/sslh/yrutschle
+DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/1.Information-Gathering/4.SSL
+DSKTPFLSDEST=/home/$USER/.local/share/applications
+DSKTPFL=sslh.desktop
+GITRESET () {
+	git clean -f
+	git fetch origin
+	git reset --hard origin/$BRANCH
+	git pull
+}
+GITSBMDLINIT () {
+	git submodule init
+	git submodule update --recursive
+	sudo updatedb && sudo ldconfig
+}
+
 echo "${bold}
  ____ ____  _     _   _ 
 / ___/ ___|| |   | | | |
@@ -10,51 +30,52 @@ echo "${bold}
  ___) |__) | |___|  _  |
 |____/____/|_____|_| |_|
             
+UPDATE
 ${normal}"
 
-GITREPOROOT=/opt/ITSEC/8.Tunnel/sslh/yrutschle/sslh
-GITREPOGITFILE=$GITREPOROOT/.git
-DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/1.Information-Gathering/4.SSL
-DSKTPFLSDEST=/home/$USER/.local/share/applications
-DSKTPFL=sslh.desktop
 
-
-if [ ! -d $GITREPOGITFILE ]
+if [ ! -d $GITCONFDIR ]
 
 then
 
-mkdir -p /opt/ITSEC/8.Tunnel/sslh/yrutschle
-cd /opt/ITSEC/8.Tunnel/sslh/yrutschle
-git clone https://github.com/yrutschle/sslh.git
+mkdir -p $GITCLONEDIR
+cd $GITCLONEDIR
+git clone -b $BRANCH $GITREPO
 
 else
 
-echo "repo exists"
+echo "${bold}REPO EXISTS, skip clone...${normal}"
 
 fi
 
 cd $GITREPOROOT
 
-if git diff-index --quiet HEAD --; then
-    echo "UP TO DATE"
-
-else
-
+if git checkout $BRANCH &&
+    git fetch origin $BRANCH &&
+    [ `git rev-list HEAD...origin/$BRANCH --count` != 0 ] &&
+    git merge origin/$BRANCH
+then
+    
 cd $GITREPOROOT
 sudo make uninstall
 make clean
-git clean -f
-git fetch origin
-git reset --hard origin/master
-git pull
-git submodule init
-git submodule update --recursive
-
+GITRESET
+GITSBMDLINIT
 make -j 4
 sudo make install
 rm -f $DSKTPFLSDEST/$DSKTPFL
 mkdir -p $DSKTPFLSDEST
 cp $DSKTPFLS/$DSKTPFL $DSKTPFLSDEST/$DSKTPFL
 
+echo "${bold}
+UPDATED
+${normal}"
+
+else
+
+echo "${bold}
+UP TO DATE
+${normal}"
+	
 fi
