+++ 8.Tunnel/stunnel/#stunnel-update.sh
@@ -3,6 +3,27 @@
 bold=$(tput bold)
 normal=$(tput sgr0)
 
+GITREPO=https://github.com/airtrack/stunnel.git
+BRANCH=master
+GITREPOROOT=/opt/ITSEC/8.Tunnel/stunnel/airtrack/stunnel
+GITCONFDIR=/opt/ITSEC/8.Tunnel/stunnel/airtrack/stunnel/.git
+GITCLONEDIR=/opt/ITSEC/8.Tunnel/stunnel/airtrack
+DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/8.Tunnel
+DSKTPFLSDEST=/home/$USER/.local/share/applications/8.Tunnel
+DSKTPFL=stunnel.desktop
+
+GITRESET () {
+	git clean -f
+	git fetch origin
+	git reset --hard origin/$BRANCH
+	git pull
+}
+GITSBMDLINIT () {
+	git submodule init
+	git submodule update --recursive
+	sudo updatedb && sudo ldconfig
+}
+
 echo "${bold}
  ____ _____ _   _ _   _ _   _ _____ _     
 / ___|_   _| | | | \ | | \ | | ____| |    
@@ -10,46 +31,48 @@ echo "${bold}
  ___) || | | |_| | |\  | |\  | |___| |___ 
 |____/ |_|  \___/|_| \_|_| \_|_____|_____|
              
+UPDATE
 ${normal}"
 
-GITREPOROOT=/opt/ITSEC/8.Tunnel/stunnel/airtrack/stunnel
-GITREPOGITFILE=$GITREPOROOT/.git
-DSKTPFLS=/opt/ITSEC-Install-Scripts/0.Initial/usrlcl/.local/share/applications/8.Tunnel
-DSKTPFLSDEST=/home/$USER/.local/share/applications/8.Tunnel
-DSKTPFL=stunnel.desktop
-
-if [ ! -d $GITREPOGITFILE ]
+if [ ! -d $GITCONFDIR ]
 
 then
 
-mkdir -p /opt/ITSEC/8.Tunnel/stunnel/airtrack
-cd /opt/ITSEC/8.Tunnel/stunnel/airtrack
-git clone https://github.com/airtrack/stunnel.git
+mkdir -p $GITCLONEDIR
+cd $GITCLONEDIR
+git clone -b $BRANCH $GITREPO
 
 else
 
-echo "repo exists"
+echo "${bold}REPO EXISTS, skip clone...${normal}"
 
 fi
 
 cd $GITREPOROOT
 
-if git diff-index --quiet HEAD --; then
-    echo "UP TO DATE"
-
-else
-
+if git checkout $BRANCH &&
+    git fetch origin $BRANCH &&
+    [ `git rev-list HEAD...origin/$BRANCH --count` != 0 ] &&
+    git merge origin/$BRANCH
+then
+    
 cd $GITREPOROOT
-git clean -f
-git fetch origin
-git reset --hard origin/master
-git pull
-git submodule init
-git submodule update --recursive
+GITRESET
+GITSBMDLINIT
 
 cargo build -v --release
 rm -f $DSKTPFLSDEST/$DSKTPFL
 mkdir -p $DSKTPFLSDEST
 cp $DSKTPFLS/$DSKTPFL $DSKTPFLSDEST/$DSKTPFL
 
+echo "${bold}
+UPDATED
+${normal}"
+
+else
+
+echo "${bold}
+UP TO DATE
+${normal}"
+	
 fi
